Move Game files into academy
- Changed tslint exception directory
- Changed relative utils dependency directory
