+++ src/components/academy/game/black-overlay/black-overlay.js
