+++ src/components/academy/game/dialog-manager/Text.js
