+++ src/components/academy/game/dialog-manager/avatar/avatar.js
