+++ src/components/academy/game/dialog-manager/dialog-manager.js
