+++ src/components/academy/game/external-manager/external-manager.js
