+++ src/components/academy/game/index.tsx
