+++ src/components/academy/game/location-manager/location-manager.js
