+++ src/components/academy/game/map-manager/map-manager.js
