+++ src/components/academy/game/object-manager/object-manager.js
