+++ src/components/academy/game/story-manager/story-manager.js
