+++ src/components/academy/game/story-xml-player.js
