+++ src/containers/GameContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
 import { bindActionCreators, Dispatch } from 'redux'
 
 import { saveCanvas } from '../actions/game'
-import Game, { DispatchProps, StateProps } from '../components/game'
+import Game, { DispatchProps, StateProps } from '../components/academy/game'
 import { IState } from '../reducers/states'
 
 const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (dispatch: Dispatch<any>) =>
